+++ bin/sync.sh
@@ -2,3 +2,4 @@
 
 git checkout v5-dev
 git pull osc v5-dev
+git pull origin v5-dev